requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
jinja2==3.1.4
pdfplumber==0.11.3
python-docx==1.1.2
groq==0.4.2
python-dotenv==1.0.0